refactor: separate fetching logic into useRequestPeople
